Move transaction::Hash test to tests module
