+++ zebra-chain/src/transaction/tests/vectors.rs
@@ -31,6 +31,19 @@ lazy_static! {
 }
 
 #[test]
+fn transactionhash_from_str() {
+    zebra_test::init();
+
+    let hash: Hash = "3166411bd5343e0b284a108f39a929fbbb62619784f8c6dafe520703b5b446bf"
+        .parse()
+        .unwrap();
+    assert_eq!(
+        format!("{:?}", hash),
+        r#"transaction::Hash("3166411bd5343e0b284a108f39a929fbbb62619784f8c6dafe520703b5b446bf")"#
+    );
+}
+
+#[test]
 fn librustzcash_tx_deserialize_and_round_trip() {
     zebra_test::init();
